+++ src/templates/mediaproject/src/config/routes.ts
@@ -1,0 +1,4 @@
+export enum RouteEnum {
+  home = '/',
+  start = '/start',
+}