+++ src/templates/mediaproject/src/types/window.d.ts
@@ -1,0 +1,7 @@
+import { WindowType } from '@ktsstudio/mediaproject-utils';
+
+declare global {
+  interface Window extends WindowType {
+    is_dev: boolean;
+  }
+}